Refactored screening question to have root parent
