+++ src/types/Question.ts
@@ -1,0 +1,3 @@
+export const ScreeningQuestionType = {
+  DROPDOWN: "dropdown",
+};
